// client/types/events.ts
export enum WSEvent {
  // Client -> Server
  CREATE_ROOM = 'CREATE_ROOM',
  JOIN_ROOM = 'JOIN_ROOM',
  LEAVE_ROOM = 'LEAVE_ROOM',
  SET_READY = 'SET_READY',
  START_GAME = 'START_GAME',
  TURN_DONE = 'TURN_DONE',
  CAST_VOTE = 'CAST_VOTE',
  SEND_CHAT = 'SEND_CHAT',
  UPDATE_SETTINGS = 'UPDATE_SETTINGS',
  MRWHITE_GUESS = 'MRWHITE_GUESS',
  RECONNECT = 'RECONNECT',
  SUBMIT_DESCRIPTION = 'SUBMIT_DESCRIPTION',
  LIST_PUBLIC_ROOMS = 'LIST_PUBLIC_ROOMS',

  // Server -> Client
  ROOM_CREATED = 'ROOM_CREATED',
  ROOM_JOINED = 'ROOM_JOINED',
  ROOM_UPDATED = 'ROOM_UPDATED',
  PUBLIC_ROOMS_LIST = 'PUBLIC_ROOMS_LIST',
  PLAYER_JOINED = 'PLAYER_JOINED',
  PLAYER_LEFT = 'PLAYER_LEFT',
  PLAYER_READY = 'PLAYER_READY',
  GAME_STARTING = 'GAME_STARTING',
  ROLE_ASSIGNED = 'ROLE_ASSIGNED',
  TURN_STARTED = 'TURN_STARTED',
  TURN_ENDED = 'TURN_ENDED',
  ROUND_ENDED = 'ROUND_ENDED',
  DISCUSSION_STARTED = 'DISCUSSION_STARTED',
  VOTE_STARTED = 'VOTE_STARTED',
  VOTE_CAST = 'VOTE_CAST',
  VOTE_RESULT = 'VOTE_RESULT',
  GAME_ENDED = 'GAME_ENDED',
  CHAT_MESSAGE = 'CHAT_MESSAGE',
  ERROR = 'ERROR',
}

export enum ErrorCode {
  ROOM_NOT_FOUND = 'ROOM_NOT_FOUND',
  ROOM_FULL = 'ROOM_FULL',
  GAME_ALREADY_STARTED = 'GAME_ALREADY_STARTED',
  NOT_HOST = 'NOT_HOST',
  NOT_ENOUGH_PLAYERS = 'NOT_ENOUGH_PLAYERS',
  PLAYER_NOT_READY = 'PLAYER_NOT_READY',
  INVALID_VOTE = 'INVALID_VOTE',
  INVALID_PAYLOAD = 'INVALID_PAYLOAD',
  PLAYER_ALREADY_IN_ROOM = 'PLAYER_ALREADY_IN_ROOM',
}